+++ tests/test_damage_types.py
@@ -1,44 +1,58 @@
-from fastapi.testclient import TestClient
 from sqlalchemy import select
 
 from server.models import DamageType
-
-from .conftest import app
-
-client = TestClient(app)
+from .conftest import client
 
 
-def test_get_damage_types(create_damage_type, db_session):
+def test_no_auth_damage_types():
     response = client.get("/api/damage_types")
+
+    assert response.status_code == 401
+    assert response.json() == {"detail": "Not authenticated"}
+
+
+def test_get_damage_types(login, create_damage_type):
+    token = login.get(name="user_token")
+    response = client.get("/api/damage_types", headers={"Authorization": f"Bearer {token}"})
+
     assert response.status_code == 200
     assert response.json() == [{"id": 1, "name": "Fire"}]
 
 
-def test_get_no_users(db_session):
-    response = client.get("/api/damage_types")
+def test_get_no_users(login):
+    token = login.get(name="user_token")
+    response = client.get("/api/damage_types", headers={"Authorization": f"Bearer {token}"})
+
     assert response.status_code == 200
     assert response.json() == []
 
 
-def test_get_damage_type(create_damage_type, db_session):
-    response = client.get("/api/damage_types/1")
+def test_get_damage_type(login, create_damage_type):
+    token = login.get(name="user_token")
+    response = client.get("/api/damage_types/1", headers={"Authorization": f"Bearer {token}"})
+
     assert response.status_code == 200
     assert response.json() == {"id": 1, "name": "Fire"}
 
 
-def test_get_no_damage_type(create_damage_type, db_session):
-    response = client.get("/api/damage_types/2")
+def test_get_no_damage_type(login, create_damage_type):
+    token = login.get(name="user_token")
+    response = client.get("/api/damage_types/2", headers={"Authorization": f"Bearer {token}"})
+
     assert response.status_code == 404
     assert response.json() == {"detail": "Damage type not found."}
 
 
-def test_post_damage_type(db_session):
+def test_post_damage_type(login):
+    token = login.get(name="user_token")
     response = client.post(
         "/api/damage_types",
+        headers={"Authorization": f"Bearer {token}"},
         json={
             "damage_type_name": "Fire",
         },
     )
+
     assert response.status_code == 201
     assert response.json() == {
         "message": "New damage type 'Fire' has been added to the database.",
@@ -46,30 +60,38 @@ def test_post_damage_type(db_session):
     }
 
 
-def test_post_duplicate_damage_type(db_session):
+def test_post_duplicate_damage_type(login):
+    token = login.get(name="user_token")
     client.post(
         "/api/damage_types",
+        headers={"Authorization": f"Bearer {token}"},
         json={
             "damage_type_name": "Fire",
         },
     )
     response = client.post(
         "/api/damage_types",
+        headers={"Authorization": f"Bearer {token}"},
         json={
             "damage_type_name": "Fire",
         },
     )
+
     assert response.status_code == 400
     assert response.json() == {"detail": "Damage type already exists."}
 
 
-def test_damage_type_put(create_damage_type, db_session):
+def test_damage_type_put(login, create_damage_type, db_session):
+    token = login.get(name="user_token")
     response = client.put(
         f"/api/damage_types/{create_damage_type.id}",
+        headers={"Authorization": f"Bearer {token}"},
         json={"damage_type_name": "Slashing"},
     )
+
     stmt = select(DamageType)
     damage_type = db_session.execute(stmt).scalar_one_or_none()
+
     assert response.status_code == 200
     assert damage_type.name == "Slashing"
     assert response.json() == {
@@ -78,33 +100,42 @@ def test_damage_type_put(create_damage_type, db_session):
     }
 
 
-def test_damage_type_duplicate_name_put(create_damage_type, db_session):
+def test_damage_type_duplicate_name_put(login, create_damage_type, db_session):
     damage_type = DamageType(name="Slashing")
     db_session.add(damage_type)
     db_session.commit()
+
+    token = login.get(name="user_token")
     response = client.put(
         f"/api/damage_types/{damage_type.id}",
+        headers={"Authorization": f"Bearer {token}"},
         json={"damage_type_name": "Fire"},
     )
+
     assert response.status_code == 400
     assert response.json() == {
         "detail": "The name you are trying to use already exists.",
     }
 
 
-def test_damage_type_fake_damage_type_put(create_race, create_damage_type, db_session):
+def test_damage_type_fake_damage_type_put(login, create_race, create_damage_type):
+    token = login.get(name="user_token")
     response = client.put(
         "/api/damage_types/2",
+        headers={"Authorization": f"Bearer {token}"},
         json={"damage_type_name": "Slashing"},
     )
+
     assert response.status_code == 404
     assert response.json() == {
         "detail": "The damage type you are trying to update does not exist.",
     }
 
 
-def test_damage_type_delete(create_damage_type, db_session):
-    response = client.delete(f"/api/damage_types/{create_damage_type.id}")
+def test_damage_type_delete(login, create_damage_type, db_session):
+    token = login.get(name="user_token")
+    response = client.delete(f"/api/damage_types/{create_damage_type.id}", headers={"Authorization": f"Bearer {token}"})
+
     damage_type = db_session.get(DamageType, create_damage_type.id)
 
     assert response.status_code == 200
@@ -112,8 +143,10 @@ def test_damage_type_delete(create_damage_type, db_session):
     assert damage_type == None
 
 
-def test_damage_type_fake_delete(create_damage_type, db_session):
-    response = client.delete(f"/api/damage_types/2")
+def test_damage_type_fake_delete(login, create_damage_type):
+    token = login.get(name="user_token")
+    response = client.delete(f"/api/damage_types/2", headers={"Authorization": f"Bearer {token}"})
+
     assert response.status_code == 404
     assert response.json() == {
         "detail": "The damage type you are trying to delete does not exist."
